Copied from new timepicker and unified component branch
