+++ packages/grafana-ui/src/utils/validate.ts
@@ -1,15 +1,24 @@
-import { EventsWithValidation, ValidationEvents, ValidationRule } from '../types';
+import { ValidationRule, ValidationEvents } from '../types/input';
+
+export enum EventsWithValidation {
+  onBlur = 'onBlur',
+  onFocus = 'onFocus',
+  onChange = 'onChange',
+}
 
 export const validate = (value: string, validationRules: ValidationRule[]) => {
-  const errors = validationRules.reduce((acc, currentRule) => {
-    if (!currentRule.rule(value)) {
-      return acc.concat(currentRule.errorMessage);
-    }
-    return acc;
-  }, []);
+  const errors = validationRules.reduce(
+    (acc, currRule) => {
+      if (!currRule.rule(value)) {
+        return acc.concat(currRule.errorMessage);
+      }
+      return acc;
+    },
+    [] as string[]
+  );
   return errors.length > 0 ? errors : null;
 };
 
-export const hasValidationEvent = (event: EventsWithValidation, validationEvents?: ValidationEvents) => {
+export const hasValidationEvent = (event: EventsWithValidation, validationEvents: ValidationEvents | undefined) => {
   return validationEvents && validationEvents[event];
 };
