--- public/app/core/components/Form/index.ts
@@ -1,1 +1,0 @@
-export { Input } from './Input';
